requirements.txt
asttokens==2.4.1
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.7
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.13.1
fonttools==4.44.3
frozendict==2.3.8
greenlet==3.0.1
html5lib==1.1
identify==2.5.32
idna==3.4
iniconfig==2.0.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
multitasking==0.0.11
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
nodeenv==1.8.0
numpy==1.26.2
packaging==23.2
pandarallel==1.6.5
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
peewee==3.17.0
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
pre-commit==3.5.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
pyarrow==14.0.1
Pygments==2.17.1
pyparsing==3.1.1
pytest==7.4.3
pytest-mock==3.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.31.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.1
scipy==1.11.4
seaborn==0.13.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
statsmodels==0.14.0
tinycss2==1.2.1
tornado==6.3.3
traitlets==5.13.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
virtualenv==20.24.6
wcwidth==0.2.10
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
yfinance==0.2.32
